apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.10.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pgvector==0.2.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sentry-sdk==1.40.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
ruff==0.1.15
mypy==1.8.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
